Adding support for updating certs with own
I've been thinking about how best to handle this to prevent foot-gun with
private key material, and keeping CertBox away from disk is likely the safest
option here, instead I'll use config or the Server to deal with sourcing
previous key material. I should later also update the tests to have a
pre-generated test key that loads and then expand testing to actually validate
the signature and not just the length of it.