Adding nesting tag support
This also now changes the interface, returning a map. Not yet doing any
transformations on the values or validation of the values which will come later.

However, for my notes it appears that roughenough's not very compliant to the
specification, and I've started leaving notes. For example, both draft -05 and
-07 say that VER is required however in my sample I don't see it. I may make
some implementation notes and raise bugs; for the time being I should be
"Postel" about it and just minimise such validation in both application code and
in tests noting where and why.

Also renaming the fixtures because of this.