Fixing Incendium runner
By putting in `runtime: false` with the dependency, a Supervisor related to
Benchee doesn't start and the benchmark fails to complete. This just fires it up
before we run out benchmark as enabling runtime will bring the kitchen sink of
Phoenix into the application in all other circumstances which is unideal.