Making recursion optional
This should make signing a bit easier - however on the flip side it will only
be a blanket all or nothing, which means if I disable it, I have to manually
then call parse on every structure, and in places we're two deep. There's some
thinking to be done how to find a smart compromise, one where we traverse the
structure, but can validate as we go.

In case you're wondering `generate_message/1` doesn't need anything done,
because I had the foresight to include a guard in the conditional to check if a
tag is recursive, but the value for it is not a map, we do nothing.